move reverse_expand weighted functionality to own file
